+++ sync-for-payables/Codat/Sync/Payables/Models/Components/CompanyRequestBody.cs
@@ -12,7 +12,6 @@ namespace Codat.Sync.Payables.Models.Components
     using Codat.Sync.Payables.Models.Components;
     using Codat.Sync.Payables.Utils;
     using Newtonsoft.Json;
-    using System.Collections.Generic;
     
     public class CompanyRequestBody
     {
@@ -30,9 +29,9 @@ namespace Codat.Sync.Payables.Models.Components
         public string? Description { get; set; }
 
         /// <summary>
-        /// Reference to the groups that the company is assigned to.
+        /// A collection of user-defined key-value pairs that store custom metadata against the company.
         /// </summary>
-        [JsonProperty("groups")]
-        public List<Items>? Groups { get; set; }
+        [JsonProperty("tags")]
+        public Tags? Tags { get; set; }
     }
 }