+++ sync-for-payables/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [csharp v4.0.0] sync-for-payables
 ### Releases
 - [NuGet v4.0.0] https://www.nuget.org/packages/Codat.Sync.Payables/4.0.0 - sync-for-payables
+
+## 2024-10-02 11:20:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.405.8 (2.429.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v5.0.0] sync-for-payables
+### Releases
+- [NuGet v5.0.0] https://www.nuget.org/packages/Codat.Sync.Payables/5.0.0 - sync-for-payables